work in progress, test cases failing...
