+++ presto-iceberg/src/main/java/io/prestosql/plugin/iceberg/IcebergMetadata.java
@@ -21,14 +21,7 @@ import io.airlift.slice.Slice;
 import io.prestosql.plugin.base.classloader.ClassLoaderSafeSystemTable;
 import io.prestosql.plugin.hive.HdfsEnvironment;
 import io.prestosql.plugin.hive.HdfsEnvironment.HdfsContext;
-import io.prestosql.plugin.hive.HiveSchemaProperties;
 import io.prestosql.plugin.hive.HiveWrittenPartitions;
-import io.prestosql.plugin.hive.TableAlreadyExistsException;
-import io.prestosql.plugin.hive.authentication.HiveIdentity;
-import io.prestosql.plugin.hive.metastore.Database;
-import io.prestosql.plugin.hive.metastore.HiveMetastore;
-import io.prestosql.plugin.hive.metastore.HivePrincipal;
-import io.prestosql.plugin.hive.metastore.Table;
 import io.prestosql.spi.PrestoException;
 import io.prestosql.spi.connector.CatalogSchemaName;
 import io.prestosql.spi.connector.ColumnHandle;
@@ -44,7 +37,6 @@ import io.prestosql.spi.connector.ConnectorTableMetadata;
 import io.prestosql.spi.connector.ConnectorTableProperties;
 import io.prestosql.spi.connector.Constraint;
 import io.prestosql.spi.connector.ConstraintApplicationResult;
-import io.prestosql.spi.connector.SchemaNotFoundException;
 import io.prestosql.spi.connector.SchemaTableName;
 import io.prestosql.spi.connector.SchemaTablePrefix;
 import io.prestosql.spi.connector.SystemTable;
@@ -58,29 +50,20 @@ import io.prestosql.spi.type.TypeManager;
 import org.apache.hadoop.fs.Path;
 import org.apache.iceberg.AppendFiles;
 import org.apache.iceberg.DataFiles;
-import org.apache.iceberg.FileFormat;
 import org.apache.iceberg.PartitionField;
-import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.PartitionSpecParser;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.SchemaParser;
 import org.apache.iceberg.Snapshot;
-import org.apache.iceberg.TableMetadata;
-import org.apache.iceberg.TableOperations;
+import org.apache.iceberg.Table;
 import org.apache.iceberg.Transaction;
-import org.apache.iceberg.catalog.Namespace;
-import org.apache.iceberg.catalog.TableIdentifier;
-import org.apache.iceberg.exceptions.NoSuchNamespaceException;
-import org.apache.iceberg.hadoop.HadoopCatalog;
 import org.apache.iceberg.types.Type;
 import org.apache.iceberg.types.TypeUtil;
 import org.apache.iceberg.types.Types;
 import org.apache.iceberg.types.Types.NestedField;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -94,241 +77,155 @@ import static com.google.common.base.Verify.verify;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
-import static io.prestosql.plugin.hive.HiveMetadata.TABLE_COMMENT;
-import static io.prestosql.plugin.hive.util.HiveWriteUtils.getTableDefaultLocation;
 import static io.prestosql.plugin.iceberg.ExpressionConverter.toIcebergExpression;
-import static io.prestosql.plugin.iceberg.IcebergSchemaProperties.getSchemaLocation;
 import static io.prestosql.plugin.iceberg.IcebergTableProperties.FILE_FORMAT_PROPERTY;
 import static io.prestosql.plugin.iceberg.IcebergTableProperties.PARTITIONING_PROPERTY;
-import static io.prestosql.plugin.iceberg.IcebergTableProperties.getFileFormat;
-import static io.prestosql.plugin.iceberg.IcebergTableProperties.getPartitioning;
-import static io.prestosql.plugin.iceberg.IcebergTableProperties.getTableLocation;
 import static io.prestosql.plugin.iceberg.IcebergUtil.getColumns;
 import static io.prestosql.plugin.iceberg.IcebergUtil.getDataPath;
 import static io.prestosql.plugin.iceberg.IcebergUtil.getFileFormat;
-import static io.prestosql.plugin.iceberg.IcebergUtil.getHadoopCatalog;
-import static io.prestosql.plugin.iceberg.IcebergUtil.getIcebergTable;
 import static io.prestosql.plugin.iceberg.IcebergUtil.getTableComment;
-import static io.prestosql.plugin.iceberg.IcebergUtil.icebergSchemaExists;
-import static io.prestosql.plugin.iceberg.IcebergUtil.icebergTableExists;
-import static io.prestosql.plugin.iceberg.IcebergUtil.isIcebergTable;
-import static io.prestosql.plugin.iceberg.IcebergUtil.useMetastore;
-import static io.prestosql.plugin.iceberg.PartitionFields.parsePartitionFields;
 import static io.prestosql.plugin.iceberg.PartitionFields.toPartitionFields;
 import static io.prestosql.plugin.iceberg.TableType.DATA;
 import static io.prestosql.plugin.iceberg.TypeConverter.toIcebergType;
 import static io.prestosql.plugin.iceberg.TypeConverter.toPrestoType;
-import static io.prestosql.spi.StandardErrorCode.INVALID_SCHEMA_PROPERTY;
 import static io.prestosql.spi.StandardErrorCode.NOT_SUPPORTED;
-import static io.prestosql.spi.StandardErrorCode.SCHEMA_NOT_EMPTY;
 import static io.prestosql.spi.type.BigintType.BIGINT;
 import static java.util.Collections.singletonList;
 import static java.util.Objects.requireNonNull;
 import static java.util.function.Function.identity;
-import static java.util.stream.Collectors.toList;
-import static org.apache.iceberg.BaseMetastoreTableOperations.ICEBERG_TABLE_TYPE_VALUE;
-import static org.apache.iceberg.BaseMetastoreTableOperations.TABLE_TYPE_PROP;
-import static org.apache.iceberg.TableMetadata.newTableMetadata;
-import static org.apache.iceberg.TableProperties.DEFAULT_FILE_FORMAT;
-import static org.apache.iceberg.Transactions.createTableTransaction;
 
-public class IcebergMetadata
+public abstract class IcebergMetadata
         implements ConnectorMetadata
 {
-    private final HiveMetastore metastore;
-    private final HdfsEnvironment hdfsEnvironment;
-    private final TypeManager typeManager;
+    protected final HdfsEnvironment hdfsEnvironment;
+    protected final TypeManager typeManager;
     private final JsonCodec<CommitTaskData> commitTaskCodec;
 
     private final Map<String, Optional<Long>> snapshotIds = new ConcurrentHashMap<>();
 
-    private Transaction transaction;
+    protected Transaction transaction;
 
     public IcebergMetadata(
-            HiveMetastore metastore,
             HdfsEnvironment hdfsEnvironment,
             TypeManager typeManager,
             JsonCodec<CommitTaskData> commitTaskCodec)
     {
-        this.metastore = requireNonNull(metastore, "metastore is null");
         this.hdfsEnvironment = requireNonNull(hdfsEnvironment, "hdfsEnvironment is null");
         this.typeManager = requireNonNull(typeManager, "typeManager is null");
         this.commitTaskCodec = requireNonNull(commitTaskCodec, "commitTaskCodec is null");
     }
 
-    @Override
-    public List<String> listSchemaNames(ConnectorSession session)
+    public static List<IcebergColumnHandle> getColumns(Schema schema, TypeManager typeManager)
     {
-        if (useMetastore(session)) {
-            return metastore.getAllDatabases();
-        }
-        else {
-            HadoopCatalog hadoopCatalog = getHadoopCatalog(hdfsEnvironment, session);
-            try {
-                List<Namespace> namespaces = hadoopCatalog.listNamespaces(Namespace.empty());
-                return namespaces.stream()
-                        .map(namespace -> namespace.toString())
-                        .collect(toList());
-            }
-            catch (NoSuchNamespaceException e) {
-                return ImmutableList.of();
-            }
-        }
+        return schema.columns().stream()
+                .map(column -> new IcebergColumnHandle(
+                        column.fieldId(),
+                        column.name(),
+                        toPrestoType(column.type(), typeManager),
+                        Optional.ofNullable(column.doc())))
+                .collect(toImmutableList());
     }
 
-    @Override
-    public Map<String, Object> getSchemaProperties(ConnectorSession session, CatalogSchemaName schemaName)
-    {
-        Optional<Database> db = metastore.getDatabase(schemaName.getSchemaName());
-        if (db.isPresent()) {
-            return HiveSchemaProperties.fromDatabase(db.get());
-        }
-
-        throw new SchemaNotFoundException(schemaName.getSchemaName());
-    }
+    public abstract Table getIcebergTable(ConnectorSession session, SchemaTableName schemaTableName) throws TableNotFoundException, UnknownTableTypeException;
 
     @Override
-    public Optional<PrestoPrincipal> getSchemaOwner(ConnectorSession session, CatalogSchemaName schemaName)
-    {
-        Optional<Database> database = metastore.getDatabase(schemaName.getSchemaName());
-        if (database.isPresent()) {
-            return database.flatMap(db -> Optional.of(new PrestoPrincipal(db.getOwnerType(), db.getOwnerName())));
-        }
-
-        throw new SchemaNotFoundException(schemaName.getSchemaName());
-    }
+    public abstract List<String> listSchemaNames(ConnectorSession session);
 
     @Override
-    public IcebergTableHandle getTableHandle(ConnectorSession session, SchemaTableName tableName)
+    public abstract  Map<String, Object> getSchemaProperties(ConnectorSession session, CatalogSchemaName schemaName);
+
+    @Override
+    public abstract Optional<PrestoPrincipal> getSchemaOwner(ConnectorSession session, CatalogSchemaName schemaName);
+
+    @Override
+    public final IcebergTableHandle getTableHandle(ConnectorSession session, SchemaTableName tableName)
     {
         IcebergTableName name = IcebergTableName.from(tableName.getTableName());
         verify(name.getTableType() == DATA, "Wrong table type: " + name.getTableType());
 
-        org.apache.iceberg.Table table;
-        if (useMetastore(session)) {
-            Optional<Table> hiveTable = metastore.getTable(new HiveIdentity(session), tableName.getSchemaName(), name.getTableName());
-            if (hiveTable.isEmpty()) {
-                return null;
-            }
-            if (!isIcebergTable(hiveTable.get())) {
-                throw new UnknownTableTypeException(tableName);
-            }
-            table = getIcebergTable(metastore, hdfsEnvironment, session, hiveTable.get().getSchemaTableName());
+        try {
+            org.apache.iceberg.Table table = getIcebergTable(session, tableName);
+            Optional<Long> snapshotId = getSnapshotId(table, name.getSnapshotId());
+
+            return new IcebergTableHandle(
+                    tableName.getSchemaName(),
+                    name.getTableName(),
+                    name.getTableType(),
+                    snapshotId,
+                    TupleDomain.all());
         }
-        else {
-            if (!icebergTableExists(hdfsEnvironment, session, tableName)) {
-                return null;
-            }
-            table = getIcebergTable(metastore, hdfsEnvironment, session, tableName);
+        catch(TableNotFoundException | UnknownTableTypeException e){
+            return null;
         }
 
-        Optional<Long> snapshotId = getSnapshotId(table, name.getSnapshotId());
-
-        return new IcebergTableHandle(
-                tableName.getSchemaName(),
-                name.getTableName(),
-                name.getTableType(),
-                snapshotId,
-                TupleDomain.all());
     }
 
     @Override
-    public Optional<SystemTable> getSystemTable(ConnectorSession session, SchemaTableName tableName)
+    public final Optional<SystemTable> getSystemTable(ConnectorSession session, SchemaTableName tableName)
     {
         return getRawSystemTable(session, tableName)
                 .map(systemTable -> new ClassLoaderSafeSystemTable(systemTable, getClass().getClassLoader()));
     }
 
-    private Optional<SystemTable> getRawSystemTable(ConnectorSession session, SchemaTableName tableName)
+    protected final Optional<SystemTable> getRawSystemTable(ConnectorSession session, SchemaTableName tableName)
     {
         IcebergTableName name = IcebergTableName.from(tableName.getTableName());
 
-        org.apache.iceberg.Table table;
+        try {
+            org.apache.iceberg.Table table = getIcebergTable(session, tableName);
 
-        if (useMetastore(session)) {
-            Optional<Table> hiveTable = metastore.getTable(new HiveIdentity(session), tableName.getSchemaName(), name.getTableName());
-            if (hiveTable.isEmpty() || !isIcebergTable(hiveTable.get())) {
-                return Optional.empty();
+            SchemaTableName systemTableName = new SchemaTableName(tableName.getSchemaName(), name.getTableNameWithType());
+            switch (name.getTableType()) {
+                case HISTORY:
+                    if (name.getSnapshotId().isPresent()) {
+                        throw new PrestoException(NOT_SUPPORTED, "Snapshot ID not supported for history table: " + systemTableName);
+                    }
+                    return Optional.of(new HistoryTable(systemTableName, table));
+                case SNAPSHOTS:
+                    if (name.getSnapshotId().isPresent()) {
+                        throw new PrestoException(NOT_SUPPORTED, "Snapshot ID not supported for snapshots table: " + systemTableName);
+                    }
+                    return Optional.of(new SnapshotsTable(systemTableName, typeManager, table));
+                case PARTITIONS:
+                    return Optional.of(new PartitionTable(systemTableName, typeManager, table, getSnapshotId(table, name.getSnapshotId())));
+                case MANIFESTS:
+                    return Optional.of(new ManifestsTable(systemTableName, table, getSnapshotId(table, name.getSnapshotId())));
+                case FILES:
+                    return Optional.of(new FilesTable(systemTableName, typeManager, table, getSnapshotId(table, name.getSnapshotId())));
             }
-            table = getIcebergTable(metastore, hdfsEnvironment, session, hiveTable.get().getSchemaTableName());
+            return Optional.empty();
         }
-        else {
-            if (!icebergTableExists(hdfsEnvironment, session, tableName)) {
-                return Optional.empty();
-            }
-            table = getIcebergTable(metastore, hdfsEnvironment, session, tableName);
+        catch (TableNotFoundException | UnknownTableTypeException e){
+            return Optional.empty();
         }
-
-        SchemaTableName systemTableName = new SchemaTableName(tableName.getSchemaName(), name.getTableNameWithType());
-        switch (name.getTableType()) {
-            case HISTORY:
-                if (name.getSnapshotId().isPresent()) {
-                    throw new PrestoException(NOT_SUPPORTED, "Snapshot ID not supported for history table: " + systemTableName);
-                }
-                return Optional.of(new HistoryTable(systemTableName, table));
-            case SNAPSHOTS:
-                if (name.getSnapshotId().isPresent()) {
-                    throw new PrestoException(NOT_SUPPORTED, "Snapshot ID not supported for snapshots table: " + systemTableName);
-                }
-                return Optional.of(new SnapshotsTable(systemTableName, typeManager, table));
-            case PARTITIONS:
-                return Optional.of(new PartitionTable(systemTableName, typeManager, table, getSnapshotId(table, name.getSnapshotId())));
-            case MANIFESTS:
-                return Optional.of(new ManifestsTable(systemTableName, table, getSnapshotId(table, name.getSnapshotId())));
-            case FILES:
-                return Optional.of(new FilesTable(systemTableName, typeManager, table, getSnapshotId(table, name.getSnapshotId())));
-        }
-        return Optional.empty();
     }
 
     @Override
-    public ConnectorTableProperties getTableProperties(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final ConnectorTableProperties getTableProperties(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         return new ConnectorTableProperties();
     }
 
     @Override
-    public ConnectorTableMetadata getTableMetadata(ConnectorSession session, ConnectorTableHandle table)
+    public final ConnectorTableMetadata getTableMetadata(ConnectorSession session, ConnectorTableHandle table)
     {
-        return getTableMetadata(session, ((IcebergTableHandle) table).getSchemaTableName());
+        return getIcebergTableMetadata(session, ((IcebergTableHandle) table).getSchemaTableName());
     }
 
     @Override
-    public List<SchemaTableName> listTables(ConnectorSession session, Optional<String> schemaName)
-    {
-        if (useMetastore(session)) {
-            return schemaName.map(Collections::singletonList)
-                    .orElseGet(metastore::getAllDatabases)
-                    .stream()
-                    .flatMap(schema -> metastore.getTablesWithParameter(schema, TABLE_TYPE_PROP, ICEBERG_TABLE_TYPE_VALUE).stream()
-                            .map(table -> new SchemaTableName(schema, table))
-                            .collect(toList())
-                            .stream())
-                    .collect(toList());
-        }
-        else {
-            HadoopCatalog hadoopCatalog = getHadoopCatalog(hdfsEnvironment, session);
-            List<Namespace> namespaces = hadoopCatalog.listNamespaces(Namespace.empty());
-            return namespaces.stream()
-                    .flatMap(namespace -> hadoopCatalog.listTables(namespace).stream()
-                            .map(tableIdentifier -> new SchemaTableName(tableIdentifier.namespace().toString(), tableIdentifier.name()))
-                            .collect(toList())
-                            .stream())
-                    .collect(toList());
-        }
-    }
+    public abstract List<SchemaTableName> listTables(ConnectorSession session, Optional<String> schemaName);
 
     @Override
-    public Map<String, ColumnHandle> getColumnHandles(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final Map<String, ColumnHandle> getColumnHandles(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         IcebergTableHandle table = (IcebergTableHandle) tableHandle;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, table.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, table.getSchemaTableName());
         return getColumns(icebergTable.schema(), typeManager).stream()
                 .collect(toImmutableMap(IcebergColumnHandle::getName, identity()));
     }
 
     @Override
-    public ColumnMetadata getColumnMetadata(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle columnHandle)
+    public final ColumnMetadata getColumnMetadata(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle columnHandle)
     {
         IcebergColumnHandle column = (IcebergColumnHandle) columnHandle;
         return ColumnMetadata.builder()
@@ -339,7 +236,7 @@ public class IcebergMetadata
     }
 
     @Override
-    public Map<SchemaTableName, List<ColumnMetadata>> listTableColumns(ConnectorSession session, SchemaTablePrefix prefix)
+    public final Map<SchemaTableName, List<ColumnMetadata>> listTableColumns(ConnectorSession session, SchemaTablePrefix prefix)
     {
         List<SchemaTableName> tables = prefix.getTable()
                 .map(ignored -> singletonList(prefix.toSchemaTableName()))
@@ -348,7 +245,7 @@ public class IcebergMetadata
         ImmutableMap.Builder<SchemaTableName, List<ColumnMetadata>> columns = ImmutableMap.builder();
         for (SchemaTableName table : tables) {
             try {
-                columns.put(table, getTableMetadata(session, table).getColumns());
+                columns.put(table, getIcebergTableMetadata(session, table).getColumns());
             }
             catch (TableNotFoundException e) {
                 // table disappeared during listing operation
@@ -361,188 +258,41 @@ public class IcebergMetadata
     }
 
     @Override
-    public void createSchema(ConnectorSession session, String schemaName, Map<String, Object> properties, PrestoPrincipal owner)
-    {
-        Optional<String> location = getSchemaLocation(properties).map(uri -> {
-            try {
-                hdfsEnvironment.getFileSystem(new HdfsContext(session, schemaName), new Path(uri));
-            }
-            catch (IOException | IllegalArgumentException e) {
-                throw new PrestoException(INVALID_SCHEMA_PROPERTY, "Invalid location URI: " + uri, e);
-            }
-            return uri;
-        });
-
-        if (useMetastore(session)) {
-            Database database = Database.builder()
-                    .setDatabaseName(schemaName)
-                    .setLocation(location)
-                    .setOwnerType(owner.getType())
-                    .setOwnerName(owner.getName())
-                    .build();
-
-            metastore.createDatabase(new HiveIdentity(session), database);
-        }
-        else {
-            if (location.isEmpty()) {
-                HadoopCatalog hadoopCatalog = getHadoopCatalog(hdfsEnvironment, session);
-                hadoopCatalog.createNamespace(Namespace.of(schemaName));
-            }
-            else {
-                HadoopCatalog hadoopCatalog = getHadoopCatalog(hdfsEnvironment, session, location.get());
-                hadoopCatalog.createNamespace(Namespace.of(schemaName));
-            }
-        }
-    }
+    public abstract void createSchema(ConnectorSession session, String schemaName, Map<String, Object> properties, PrestoPrincipal owner);
 
     @Override
-    public void dropSchema(ConnectorSession session, String schemaName)
-    {
-        if (useMetastore(session)) {
-            // basic sanity check to provide a better error message
-            if (!listTables(session, Optional.of(schemaName)).isEmpty() ||
-                    !listViews(session, Optional.of(schemaName)).isEmpty()) {
-                throw new PrestoException(SCHEMA_NOT_EMPTY, "Schema not empty: " + schemaName);
-            }
-            metastore.dropDatabase(new HiveIdentity(session), schemaName);
-        }
-        else {
-            HadoopCatalog hadoopCatalog = getHadoopCatalog(hdfsEnvironment, session);
-            hadoopCatalog.dropNamespace(Namespace.of(schemaName));
-        }
-    }
+    public abstract void dropSchema(ConnectorSession session, String schemaName);
 
     @Override
-    public void renameSchema(ConnectorSession session, String source, String target)
-    {
-        metastore.renameDatabase(new HiveIdentity(session), source, target);
-    }
+    public abstract void renameSchema(ConnectorSession session, String source, String target);
 
     @Override
-    public void setSchemaAuthorization(ConnectorSession session, String source, PrestoPrincipal principal)
-    {
-        metastore.setDatabaseOwner(new HiveIdentity(session), source, HivePrincipal.from(principal));
-    }
+    public abstract void setSchemaAuthorization(ConnectorSession session, String source, PrestoPrincipal principal);
 
     @Override
-    public void createTable(ConnectorSession session, ConnectorTableMetadata tableMetadata, boolean ignoreExisting)
+    public final void createTable(ConnectorSession session, ConnectorTableMetadata tableMetadata, boolean ignoreExisting)
     {
         Optional<ConnectorNewTableLayout> layout = getNewTableLayout(session, tableMetadata);
         finishCreateTable(session, beginCreateTable(session, tableMetadata, layout), ImmutableList.of(), ImmutableList.of());
     }
 
     @Override
-    public void setTableComment(ConnectorSession session, ConnectorTableHandle tableHandle, Optional<String> comment)
-    {
-        IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
-        metastore.commentTable(new HiveIdentity(session), handle.getSchemaName(), handle.getTableName(), comment);
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, handle.getSchemaTableName());
-        if (comment.isEmpty()) {
-            icebergTable.updateProperties().remove(TABLE_COMMENT).commit();
-        }
-        else {
-            icebergTable.updateProperties().set(TABLE_COMMENT, comment.get()).commit();
-        }
-    }
+    public abstract void setTableComment(ConnectorSession session, ConnectorTableHandle tableHandle, Optional<String> comment);
 
     @Override
-    public ConnectorOutputTableHandle beginCreateTable(ConnectorSession session, ConnectorTableMetadata tableMetadata, Optional<ConnectorNewTableLayout> layout)
-    {
-        SchemaTableName schemaTableName = tableMetadata.getTable();
-        String schemaName = schemaTableName.getSchemaName();
-        String tableName = schemaTableName.getTableName();
-
-        Schema schema = toIcebergSchema(tableMetadata.getColumns());
-
-        PartitionSpec partitionSpec = parsePartitionFields(schema, getPartitioning(tableMetadata.getProperties()));
-
-        if (useMetastore(session)) {
-            Database database = metastore.getDatabase(schemaName)
-                    .orElseThrow(() -> new SchemaNotFoundException(schemaName));
-
-            HdfsContext hdfsContext = new HdfsContext(session, schemaName, tableName);
-            HiveIdentity identity = new HiveIdentity(session);
-            String targetPath = getTableLocation(tableMetadata.getProperties());
-            if (targetPath == null) {
-                targetPath = getTableDefaultLocation(database, hdfsContext, hdfsEnvironment, schemaName, tableName).toString();
-            }
-
-            TableOperations operations = new HiveTableOperations(metastore, hdfsEnvironment, hdfsContext, identity, schemaName, tableName, session.getUser(), targetPath);
-            if (operations.current() != null) {
-                throw new TableAlreadyExistsException(schemaTableName);
-            }
-
-            ImmutableMap.Builder<String, String> propertiesBuilder = ImmutableMap.builderWithExpectedSize(2);
-            FileFormat fileFormat = getFileFormat(tableMetadata.getProperties());
-            propertiesBuilder.put(DEFAULT_FILE_FORMAT, fileFormat.toString());
-            if (tableMetadata.getComment().isPresent()) {
-                propertiesBuilder.put(TABLE_COMMENT, tableMetadata.getComment().get());
-            }
-
-            TableMetadata metadata = newTableMetadata(schema, partitionSpec, targetPath, propertiesBuilder.build());
-
-            transaction = createTableTransaction(tableName, operations, metadata);
-
-            return new IcebergWritableTableHandle(
-                    schemaName,
-                    tableName,
-                    SchemaParser.toJson(metadata.schema()),
-                    PartitionSpecParser.toJson(metadata.spec()),
-                    getColumns(metadata.schema(), typeManager),
-                    targetPath,
-                    fileFormat);
-        }
-        else {
-            if (!icebergSchemaExists(hdfsEnvironment, session, schemaName)) {
-                throw new SchemaNotFoundException(schemaName);
-            }
-            if (icebergTableExists(hdfsEnvironment, session, schemaTableName)) {
-                throw new TableAlreadyExistsException(schemaTableName);
-            }
-
-            ImmutableMap.Builder<String, String> propertiesBuilder = ImmutableMap.builderWithExpectedSize(2);
-            FileFormat fileFormat = getFileFormat(tableMetadata.getProperties());
-            propertiesBuilder.put(DEFAULT_FILE_FORMAT, fileFormat.toString());
-            if (tableMetadata.getComment().isPresent()) {
-                propertiesBuilder.put(TABLE_COMMENT, tableMetadata.getComment().get());
-            }
-
-            HdfsContext hdfsContext = new HdfsContext(session, schemaName, tableName);
-            String targetPath = getTableLocation(tableMetadata.getProperties());
-            HadoopCatalog catalog;
-            if (targetPath == null) {
-                catalog = getHadoopCatalog(hdfsEnvironment, session);
-            }
-            else {
-                catalog = getHadoopCatalog(hdfsEnvironment, session, targetPath);
-            }
-
-            Map<String, String> properties = propertiesBuilder.build();
-            TableIdentifier tableIdentifier = TableIdentifier.of(schemaName, tableName);
-            transaction = catalog.newCreateTableTransaction(tableIdentifier, schema, partitionSpec, properties);
-
-            return new IcebergWritableTableHandle(
-                    schemaName,
-                    tableName,
-                    SchemaParser.toJson(transaction.table().schema()),
-                    PartitionSpecParser.toJson(transaction.table().spec()),
-                    getColumns(transaction.table().schema(), typeManager),
-                    transaction.table().location(),
-                    fileFormat);
-        }
-    }
+    public abstract ConnectorOutputTableHandle beginCreateTable(ConnectorSession session, ConnectorTableMetadata tableMetadata, Optional<ConnectorNewTableLayout> layout);
 
     @Override
-    public Optional<ConnectorOutputMetadata> finishCreateTable(ConnectorSession session, ConnectorOutputTableHandle tableHandle, Collection<Slice> fragments, Collection<ComputedStatistics> computedStatistics)
+    public final Optional<ConnectorOutputMetadata> finishCreateTable(ConnectorSession session, ConnectorOutputTableHandle tableHandle, Collection<Slice> fragments, Collection<ComputedStatistics> computedStatistics)
     {
         return finishInsert(session, (IcebergWritableTableHandle) tableHandle, fragments, computedStatistics);
     }
 
     @Override
-    public ConnectorInsertTableHandle beginInsert(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final ConnectorInsertTableHandle beginInsert(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         IcebergTableHandle table = (IcebergTableHandle) tableHandle;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, table.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, table.getSchemaTableName());
 
         transaction = icebergTable.newTransaction();
 
@@ -556,8 +306,9 @@ public class IcebergMetadata
                 getFileFormat(icebergTable));
     }
 
+
     @Override
-    public Optional<ConnectorOutputMetadata> finishInsert(ConnectorSession session, ConnectorInsertTableHandle insertHandle, Collection<Slice> fragments, Collection<ComputedStatistics> computedStatistics)
+    public final Optional<ConnectorOutputMetadata> finishInsert(ConnectorSession session, ConnectorInsertTableHandle insertHandle, Collection<Slice> fragments, Collection<ComputedStatistics> computedStatistics)
     {
         IcebergWritableTableHandle table = (IcebergWritableTableHandle) insertHandle;
         org.apache.iceberg.Table icebergTable = transaction.table();
@@ -598,85 +349,53 @@ public class IcebergMetadata
     }
 
     @Override
-    public ColumnHandle getUpdateRowIdColumnHandle(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final ColumnHandle getUpdateRowIdColumnHandle(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         return new IcebergColumnHandle(0, "$row_id", BIGINT, Optional.empty());
     }
 
     @Override
-    public Optional<Object> getInfo(ConnectorTableHandle tableHandle)
+    public final Optional<Object> getInfo(ConnectorTableHandle tableHandle)
     {
         IcebergTableHandle table = (IcebergTableHandle) tableHandle;
         return Optional.of(new IcebergInputInfo(table.getSnapshotId()));
     }
 
     @Override
-    public void dropTable(ConnectorSession session, ConnectorTableHandle tableHandle)
-    {
-        IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
-        if (useMetastore(session)) {
-            metastore.dropTable(new HiveIdentity(session), handle.getSchemaName(), handle.getTableName(), true);
-        }
-        else {
-            TableIdentifier tableIdentifier = TableIdentifier.of(handle.getSchemaName(), handle.getTableName());
-            getHadoopCatalog(hdfsEnvironment, session).dropTable(tableIdentifier, false);
-        }
-    }
+    public abstract void dropTable(ConnectorSession session, ConnectorTableHandle tableHandle);
 
     @Override
-    public void renameTable(ConnectorSession session, ConnectorTableHandle tableHandle, SchemaTableName newTable)
-    {
-        IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
-        if (useMetastore(session)) {
-            metastore.renameTable(new HiveIdentity(session), handle.getSchemaName(), handle.getTableName(), newTable.getSchemaName(), newTable.getTableName());
-        }
-        else {
-            TableIdentifier tableIdentifier = TableIdentifier.of(handle.getSchemaName(), handle.getTableName());
-            TableIdentifier newTableIdentifier = TableIdentifier.of(newTable.getSchemaName(), newTable.getTableName());
-            getHadoopCatalog(hdfsEnvironment, session).renameTable(tableIdentifier, newTableIdentifier);
-        }
-    }
+    public abstract void renameTable(ConnectorSession session, ConnectorTableHandle tableHandle, SchemaTableName newTable);
 
     @Override
-    public void addColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnMetadata column)
+    public final void addColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnMetadata column)
     {
         IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, handle.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, handle.getSchemaTableName());
         icebergTable.updateSchema().addColumn(column.getName(), toIcebergType(column.getType())).commit();
     }
 
     @Override
-    public void dropColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle column)
+    public final void dropColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle column)
     {
         IcebergTableHandle icebergTableHandle = (IcebergTableHandle) tableHandle;
         IcebergColumnHandle handle = (IcebergColumnHandle) column;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, icebergTableHandle.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, icebergTableHandle.getSchemaTableName());
         icebergTable.updateSchema().deleteColumn(handle.getName()).commit();
     }
 
     @Override
-    public void renameColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle source, String target)
+    public final void renameColumn(ConnectorSession session, ConnectorTableHandle tableHandle, ColumnHandle source, String target)
     {
         IcebergTableHandle icebergTableHandle = (IcebergTableHandle) tableHandle;
         IcebergColumnHandle columnHandle = (IcebergColumnHandle) source;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, icebergTableHandle.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, icebergTableHandle.getSchemaTableName());
         icebergTable.updateSchema().renameColumn(columnHandle.getName(), target).commit();
     }
 
-    private ConnectorTableMetadata getTableMetadata(ConnectorSession session, SchemaTableName table)
+    private final ConnectorTableMetadata getIcebergTableMetadata(ConnectorSession session, SchemaTableName table) throws TableNotFoundException, UnknownTableTypeException
     {
-        if (useMetastore(session)) {
-            if (metastore.getTable(new HiveIdentity(session), table.getSchemaName(), table.getTableName()).isEmpty()) {
-                throw new TableNotFoundException(table);
-            }
-        }
-        else {
-            if (!icebergTableExists(hdfsEnvironment, session, table)) {
-                throw new TableNotFoundException(table);
-            }
-        }
-
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, table);
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, table);
 
         List<ColumnMetadata> columns = getColumnMetadatas(icebergTable);
 
@@ -689,7 +408,7 @@ public class IcebergMetadata
         return new ConnectorTableMetadata(table, columns, properties.build(), getTableComment(icebergTable));
     }
 
-    private List<ColumnMetadata> getColumnMetadatas(org.apache.iceberg.Table table)
+    private final List<ColumnMetadata> getColumnMetadatas(org.apache.iceberg.Table table)
     {
         return table.schema().columns().stream()
                 .map(column -> {
@@ -703,7 +422,7 @@ public class IcebergMetadata
                 .collect(toImmutableList());
     }
 
-    private static Schema toIcebergSchema(List<ColumnMetadata> columns)
+    protected final static Schema toIcebergSchema(List<ColumnMetadata> columns)
     {
         List<NestedField> icebergColumns = new ArrayList<>();
         for (ColumnMetadata column : columns) {
@@ -723,50 +442,50 @@ public class IcebergMetadata
     }
 
     @Override
-    public Optional<ConnectorTableHandle> applyDelete(ConnectorSession session, ConnectorTableHandle handle)
+    public final Optional<ConnectorTableHandle> applyDelete(ConnectorSession session, ConnectorTableHandle handle)
     {
         return Optional.of(handle);
     }
 
     @Override
-    public ConnectorTableHandle beginDelete(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final ConnectorTableHandle beginDelete(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         throw new PrestoException(NOT_SUPPORTED, "This connector only supports delete where one or more partitions are deleted entirely");
     }
 
     @Override
-    public OptionalLong executeDelete(ConnectorSession session, ConnectorTableHandle tableHandle)
+    public final OptionalLong executeDelete(ConnectorSession session, ConnectorTableHandle tableHandle)
     {
         IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
 
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, handle.getSchemaTableName());
+        try {
+            org.apache.iceberg.Table icebergTable = getIcebergTable(session, handle.getSchemaTableName());
 
-        icebergTable.newDelete()
-                .deleteFromRowFilter(toIcebergExpression(handle.getPredicate()))
-                .commit();
+            icebergTable.newDelete()
+                    .deleteFromRowFilter(toIcebergExpression(handle.getPredicate()))
+                    .commit();
 
-        // TODO: it should be possible to return number of deleted records
-        return OptionalLong.empty();
+            // TODO: it should be possible to return number of deleted records
+            return OptionalLong.empty();
+        }
+        catch(TableNotFoundException | UnknownTableTypeException e){
+            return OptionalLong.empty();
+        }
     }
 
     @Override
-    public boolean usesLegacyTableLayouts()
+    public final boolean usesLegacyTableLayouts()
     {
         return false;
     }
 
-    public HiveMetastore getMetastore()
-    {
-        return metastore;
-    }
-
-    public void rollback()
+    public final void rollback()
     {
         // TODO: cleanup open transaction
     }
 
     @Override
-    public Optional<ConstraintApplicationResult<ConnectorTableHandle>> applyFilter(ConnectorSession session, ConnectorTableHandle handle, Constraint constraint)
+    public final Optional<ConstraintApplicationResult<ConnectorTableHandle>> applyFilter(ConnectorSession session, ConnectorTableHandle handle, Constraint constraint)
     {
         IcebergTableHandle table = (IcebergTableHandle) handle;
 
@@ -783,7 +502,7 @@ public class IcebergMetadata
             return Optional.empty();
         }
 
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, table.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, table.getSchemaTableName());
 
         List<PartitionField> fields = icebergTable.spec().fields().stream()
                 .filter(field -> field.transform().isIdentity())
@@ -813,17 +532,18 @@ public class IcebergMetadata
     }
 
     @Override
-    public TableStatistics getTableStatistics(ConnectorSession session, ConnectorTableHandle tableHandle, Constraint constraint)
+    public final TableStatistics getTableStatistics(ConnectorSession session, ConnectorTableHandle tableHandle, Constraint constraint)
     {
         IcebergTableHandle handle = (IcebergTableHandle) tableHandle;
-        org.apache.iceberg.Table icebergTable = getIcebergTable(metastore, hdfsEnvironment, session, handle.getSchemaTableName());
+        org.apache.iceberg.Table icebergTable = getIcebergTable(session, handle.getSchemaTableName());
         return TableStatisticsMaker.getTableStatistics(typeManager, constraint, handle, icebergTable);
     }
 
-    private Optional<Long> getSnapshotId(org.apache.iceberg.Table table, Optional<Long> snapshotId)
+    protected final Optional<Long> getSnapshotId(org.apache.iceberg.Table table, Optional<Long> snapshotId)
     {
         return snapshotIds.computeIfAbsent(table.toString(), ignored -> snapshotId
                 .map(id -> IcebergUtil.resolveSnapshotId(table, id))
                 .or(() -> Optional.ofNullable(table.currentSnapshot()).map(Snapshot::snapshotId)));
     }
+
 }
