+++ presto-iceberg/src/main/java/io/prestosql/plugin/iceberg/InternalIcebergConnectorFactory.java
@@ -93,6 +93,10 @@ public final class InternalIcebergConnectorFactory
             IcebergTableProperties icebergTableProperties = injector.getInstance(IcebergTableProperties.class);
             Set<Procedure> procedures = injector.getInstance((Key<Set<Procedure>>) Key.get(Types.setOf(Procedure.class)));
 
+            //config.getOrDefault("iceberg.catalog-type", "hive")
+
+
+            //IcebergSessionProperties.getCatalogType(session);
             return new IcebergConnector(
                     lifeCycleManager,
                     transactionManager,
