+++ presto-iceberg/src/test/java/io/prestosql/plugin/iceberg/TestIcebergMetadataListing.java
@@ -122,6 +122,6 @@ public class TestIcebergMetadataListing
     public void testTableValidation()
     {
         assertQuerySucceeds("SELECT * FROM iceberg.test_schema.iceberg_table1");
-        assertQueryFails("SELECT * FROM iceberg.test_schema.hive_table", "Not an Iceberg table: test_schema.hive_table");
+        assertQueryFails("SELECT * FROM iceberg.test_schema.hive_table", "Table 'iceberg.test_schema.hive_table' does not exist");
     }
 }
